Complete project requirements: API auth/reviews/AI/dashboard, separated frontend, tests and docs

+++ app/Models/Review.php
@@ -19,7 +19,7 @@ class Review extends Model
 
     protected $casts = [
         'topics' => 'array',
-        'score'  => 'float',
+        'score'  => 'integer',
     ];
 
     public function user()
